#1: Move transformations into transactions service
