+++ src/transactions/transactions.controller.ts
@@ -1,7 +1,6 @@
 import { Body, Controller, Delete, Get, Param, Post, Put, Query, UploadedFiles, UseInterceptors } from '@nestjs/common';
 import { IsNumberString, IsString } from 'class-validator';
 import { FilesInterceptor } from '@nestjs/platform-express';
-import * as XLSX from 'xlsx';
 
 import { TransactionsService } from './transactions.service';
 import { Transaction } from './transaction.entity';
@@ -43,36 +42,7 @@ export class TransactionsController {
   @Post('batch/xlxs')
   @UseInterceptors(FilesInterceptor('upload'))
   async createManyExcel(@UploadedFiles() upload, @Query() params: ParamsCreateMany) {
-    const workBook = XLSX.read(upload[0].buffer, {cellDates: true});
-    let json: any;
-    let workSheet: XLSX.WorkSheet;
-    let formatted: CreateTransactionDto[] = [];
-    /* tslint:disable:no-string-literal */
-    if (params.format === 'skandia') {
-      workSheet = workBook.Sheets.Kontoutdrag;
-      json = XLSX.utils.sheet_to_json(workSheet);
-
-      formatted = json.map((row => {
-        return {
-          name: row['Beskrivning'],
-          date: row['Bokf. datum'],
-          amount: row['Belopp']
-        };
-      }));
-    }
-
-    if (params.format === 'norwegian') {
-      workSheet = workBook.Sheets.transactions;
-      json = XLSX.utils.sheet_to_json(workSheet);
-      formatted = json.map((row => {
-        return {
-          name: row['Text'],
-          date: new Date(row['TransactionDate']).toISOString(),
-          amount: row['Amount']
-        };
-      }));
-    }
-    return this.transactionsService.createMany(formatted);
+    return this.transactionsService.createManyExcel(upload[0].buffer, params.format);
   }
 
   @Delete(':id')
